requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.14
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.4.0
astroid==3.3.10
attrs==25.3.0
backports.tarfile==1.2.0
blinker==1.8.2
build==1.2.2.post1
CacheControl==0.14.1
certifi==2024.8.30
cffi==1.17.1
cfgv==3.4.0
charset-normalizer==3.4.1
cleo==2.1.0
click==8.2.1
crashtest==0.4.1
cryptography==44.0.0
dataclasses-json==0.6.7
dep-logic==0.4.6
dill==0.4.0
distlib==0.3.8
distro==1.9.0
dulwich==0.22.7
fastjsonschema==2.21.1
filelock==3.16.0
findpython==0.6.1
frozenlist==1.5.0
greenlet==3.1.1
h11==0.14.0
hishel==0.0.30
httpcore==1.0.5
httpx==0.27.2
httpx-sse==0.4.0
identify==2.6.0
idna==3.8
importlib_metadata==8.5.0
installer==0.7.0
isort==6.0.1
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
jeepney==0.8.0
jiter==0.9.0
jsonpatch==1.33
jsonpointer==3.0.0
keyring==25.6.0
langchain==0.3.26
langchain-community==0.3.27
langchain-core>=0.3.68,<0.4.0
langchain-text-splitters==0.3.8
langsmith==0.4.4
logmine==0.4.1
markdown-it-py==3.0.0
marshmallow==3.26.1
mccabe==0.7.0
mdurl==0.1.2
more-itertools==10.5.0
msgpack==1.1.0
multidict==6.2.0
mypy-extensions==1.0.0
nodeenv==1.9.1
numpy==2.2.4
openai==1.68.2
orjson==3.10.15
packaging==24.1
pbs-installer==2024.9.9
pdm==2.18.2
pip-tools==7.4.1
pkginfo==1.12.0
platformdirs==4.3.2
poetry==2.0.0
poetry-core==2.0.0
pre-commit==3.8.0
propcache==0.3.0
pycparser==2.22
pydantic==2.10.6
pydantic-settings==2.8.1
pydantic_core==2.27.2
Pygments==2.18.0
pylint==3.3.7
pyproject_hooks==1.1.0
python-dotenv==1.0.1
PyYAML==6.0.2
RapidFuzz==3.11.0
requests==2.32.3
requests-toolbelt==1.0.0
resolvelib==1.0.1
rich==13.8.1
SecretStorage==3.3.3
shellingham==1.5.4
slack_bolt==1.23.0
slack_sdk==3.35.0
sniffio==1.3.1
socksio==1.0.0
SQLAlchemy==2.0.39
tenacity==9.0.0
tomlkit==0.13.2
tqdm==4.67.1
trove-classifiers==2025.1.6.15
truststore==0.9.2
typing-inspect==0.9.0
typing_extensions==4.12.2
unearth==0.17.2
urllib3==2.3.0
virtualenv==20.28.1
xmltodict==0.14.2
yarl==1.18.3
zipp==3.21.0
zstandard==0.23.0
